cardlab

Struct Card

pub struct Card;

Represents a standard playing card with a suit and a rank.

Each card consists of one of the four suits (Diamond, Club, Heart, Spade), as well as a rank that can be either a number card (2-10) or a face card (Jack, Queen, King, Ace).

This type implements the traits PartialEq, Eq, PartialOrd, and Ord.

Examples

let ace_spades = Card::new("spade", 14);
let king_spades = Card::new("spade", 13);
let ten_spades = Card::new("spade", 10);
let five_spades = Card::new("spade", 5);
let ace_hearts = Card::new("heart", 14);
let ace_clubs = Card::new("club", 14);

// Compare cards of same suit but different rank.
assert!(ace_spades > king_spades);
assert!(king_spades > ten_spades);
assert!(ten_spades > five_spades);

// Compare cards of same rank but different suit.
assert!(ace_spades > ace_hearts);
assert!(ace_hearts > ace_clubs);

// Test equality between identical cards.
assert!(ace_spades == Card::new("spade", 14));

Implementations

impl Card

pub fn new(suit: &str, rank: u8) -> Self

Creates a new Card instance from a suit name and numeric rank.

Parameters

suit - A string representing the suit:

"diamond"

"club"

"heart"

"spade"

rank - A number from 1-13 representing the rank:

2-10: Number cards

11: Jack

12: Queen

13: King

14: Ace

Panics

If suit is not one of: "spade", "heart", "club", "diamond".

If rank is not in the range [2, 14].

Examples

let ace_of_spades = Card::new("spade", 14);
let two_of_hearts = Card::new("heart", 2);
let three_of_clubs = Card::new("club", 3);

pub fn suit_name(&self) -> &'static str

Returns a string representation of this card’s suit.

Examples

let card = Card::new("club", 3);
assert_eq!(card.suit_name(), "club");

pub fn rank_value(&self) -> u8

Returns the numeric rank of this card, where number cards = 2-10, and Jack = 11, Queen = 12, King = 13, and Ace = 14.

Examples

let card = Card::new("club", 3);
assert_eq!(card.rank_value(), 3);